code/requirements.txt
--index-url https://pypi.python.org/simple/
--extra-index-url https://download.pytorch.org/whl/cu118

faiss-cpu==1.7.4
torch==2.1.1+cu118
torchvision==0.16.1+cu118
transformers==4.36.0
accelerate==0.21.0
bitsandbytes==0.40.2
scikit-learn==1.4.0
matplotlib==3.8.2
tqdm==4.66.1
unstructured[pdf]==0.12.3
pytesseract==0.3.10
sentence-transformers==2.3.1
# flash-attn==2.5.2 # Fails to install if installed before or during torch install
text_generation==0.6.1
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.5
langsmith==0.0.77
jupyterlab==4.0.8
wheel==0.42.0
seaborn==0.13.2
-e .
